Rama javier a main

+++ Dominio/DetalleCompra.cs
@@ -10,7 +10,7 @@ namespace Dominio
     {
             public int Id { get; set; }
             public int Cantidad { get; set; }
-            public int PrecioUnitario { get; set; }
+            public decimal PrecioUnitario { get; set; }
             public string ProductoNombre { get; set; }
             public int IdProducto { get; set; }
             public int IdCompra { get; set; }
